move shuffle to utility module
